pyproject.toml
[project]
name = "zotan"
version = "0.1.0"
description = "Zotan AI Agent Framework"
readme = "docs/ARCHITECTURE.md"
requires-python = ">=3.13"
dependencies = [
    "cappa>=0.1.0",
    "pydantic_ai_slim[anthropic,openai]>=1.0.0",
    "tiktoken>=0.5.0",
    "tomli_w>=1.0.0",
]

[project.scripts]
zotan = "zotan.cli:main"

[project.optional-dependencies]
dev = [
    "pyright>=1.1.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = [
    "zotan",
    "zotan.spin",
    "zotan.ui",
    "zotan.tools",
]

[tool.setuptools.package-data]
zotan = ["skills/**/*"]

[tool.isort]
profile = "black"
line_length = 65535

[tool.black]
target_version = ["py313"]
line_length = 65535

[tool.pyright]
pythonVersion = "3.13"
typeCheckingMode = "strict"
include = ["zotan"]
exclude = ["zotan/skills/**"]
